pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mask-orchestrator"
version = "0.1.0"
description = "Multi-Agent Orchestrator with MASK and A2A protocol"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Colin Lee", email = "colinlee0924@gmail.com" }
]
keywords = ["mask", "orchestrator", "agent", "a2a", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mask-kernel[phoenix,anthropic] @ git+https://github.com/colinlee0924/mask-kernel.git",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "a2a-sdk>=0.3.0",
]

[project.urls]
Homepage = "https://github.com/colinlee0924/mask-orchestrator"
Repository = "https://github.com/colinlee0924/mask-orchestrator"
Issues = "https://github.com/colinlee0924/mask-orchestrator/issues"

[project.scripts]
mask-orchestrator = "mask_orchestrator.main:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mask_orchestrator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
